pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.3.4",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.13",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
